Moved player.cs to MainCamera and added player size scaling
player.cs is now attached to the MainCamera, which moves with the active player body.

The player's size can now be changed by calling ResizePlayer() and passing in the deisred change in size. Ex: 0.1 = +10%, -0.2 = -20%. Values related to movement speed and jump height will scale linearly based on the current overall size multiplier.